Refactored Trace Log for feign client
